feat: Modify gpt script to read dotfile from home directory
This commit modifies the gpt script to read the dotfile from the user's home directory instead of the current working directory. This change enables the script to be called from /usr/local/bin.

The dotfile path is now defined using the os.path.expanduser("~") method to get the path to the home directory. If the dotfile exists, it is read using the configparser module. Otherwise, a new dotfile is created with default values.

This change improves the usability and flexibility of the gpt script.