Refactor: Stop script from writing sample config file
This commit modifies the `gpt` script to stop it from writing a sample config file to the present working directory. Instead, it now prints a message to inform the user that a config file is required as `~/.openai_config`. This change improves the script's usability and prevents clutter in the working directory.